tests.plugins: Moved 'docs/tests plugins' into 'plugins_meta'
